Using centralized search category configuration from utils.
TH-1151.
